requirements.txt
cryptography~=46.0.5
pillow~=12.1.1
pip~=26.0.1
cffi~=2.0.0
Django~=6.0.2
asgiref~=3.11.1
sqlparse~=0.5.5
requests~=2.32.5
django-allauth~=65.14.3
PyJWT~=2.11.0
certifi~=2026.1.4
urllib3~=2.6.3
idna~=3.11
charset-normalizer~=3.4.4
python-dotenv~=1.2.1